De-dupe some code
