+++ bin/fuel-indexer/Cargo.toml
@@ -19,7 +19,6 @@ fuel-indexer-database = { version = "0.2", path = "../../packages/fuel-indexer-d
 fuel-indexer-lib = { version = "0.2", path = "../../packages/fuel-indexer-lib" }
 tokio = { version = "1.8", features = ["macros", "rt-multi-thread", "sync", "process"] }
 tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["ansi", "json", "env-filter"] }
 
 [features]
 default = ["api-server"]
